Try to reduce the number of I/O iterations
For halos not near the outside of the chunk read in we will often have
particles out to a much larger radius than the requested read_radius.
For each cell we compute a lower limit on the radius in which we have
all of the particles and allow the search radius for each halo in the
cell to increase to at least this value before reading in more cells.